Change the SoundDevice interface. The reason is that the previous one was using blocking I/O always, which caused the playback part to not be able to catch up (thus delaying the playback of the received packets).
So, now the playback is opened in non-blocking mode, leaving more than 
enough room for doing processing stuff. Before the write function is called,
there is some synchronization and sleep stuff, to avoid writing too fast 
to the sound device. This time is for now set to 20ms.

Changing this behavior from blocking to non-blocking also has effect on the 
actual implementation of the devices ... they all need to be fine tuned. They
all work and playback the audio, but you may experience a background noise
due to playback suffering underruns ... let me know and we can try to fix it
via the device settings (buffer sizes, triggers and so on).

BTW, the direct sound stuff ... is not implemented ... if it is useless, why is
still there?

I also moved a lot of repeated code from each device .cxx file into the sounddevice.cxx, 
hopefully creating a nicer implementation.

Another thing that needs to be tested is the actual delay ... we still need to 
implement a jitter buffer in BasicSoundSource to completely get rid of audio problems.
I am on it.

So ... the idea is this commit may temporarily leave sound playback worse than it was ... 
but it is going to be fixed immediately :D

Good night!




git-svn-id: svn://svn.minisip.org/minisip/trunk@1640 63dae60f-c5d3-0310-92a3-fcbf00261c72