3 patches to the mikey stuff ... thanks to Mikael Magnusson
* The default policy parameter should be used if the parameter wasn't transmitted
in the MIKEY SP. Implementing it in getPolicyParamTypeValue makes seance
to me. Observe that getPolicyParamType returns NULL instead of the
default value. Is a function returning the default parameter value need?

For example:

 uint8_t getDefaultParamTypeValue(uint8_t prot_type, uint8_t policy_type);

The default auth tag length should also be 10 instead of 4.

* declaring parseResponse, setOffer, buildResponse and authenticate in
keyagreement and defining them in its subclasses makes it easier to call
those functions without explicitly check and cast the keyagreement to the
correct subclass and call respective function in MikeyMessage.

* This patch adds support for custom authentication tag length in minisip.
The default authentication tag length is 10 bytes for SRTP, and smaller
values are not recommended according to RFC 3711.




git-svn-id: svn://svn.minisip.org/minisip/trunk@1685 63dae60f-c5d3-0310-92a3-fcbf00261c72