Fix by Mikael Magnusson ...
I found two problems with the TLS support, and am including a patch.
A new connection is created because a port number mismatch. The TLS
socket was always connected to port 5061 instead of the specified port,
which was used in the call to findStreamSocket.

And the proxy port was always reset to 5060 after a restart of minisip,
since the port number never was saved.




git-svn-id: svn://svn.minisip.org/minisip/trunk@1690 63dae60f-c5d3-0310-92a3-fcbf00261c72