Taaadaaaa!
Compilation in Debian Stable is re-established (not without sweat).

I added in the Documentation/build/troubleshoot a patched version of some of the 
libtool files (libtool, libtoolize and ltmain.sh). It contains a patch which fixes
the .libs/.libs problem. I bumped (articificially) the version of these files, from
1.5.6 to 1.5.6.1 (so that our configuration tools know about the patch).

I tried also building on Debian Testing and it worked fine.



git-svn-id: svn://svn.minisip.org/minisip/trunk@2566 63dae60f-c5d3-0310-92a3-fcbf00261c72